requirements.txt
aiogram>=3.4.1
openai>=1.40.0
aiohttp==3.9.5
apscheduler==3.10.4
pytz==2024.1
httpx==0.27.0
asyncpg==0.29.0
pytest==8.3.2
pytest-asyncio==0.23.8
psutil==5.9.4
google-auth==2.28.1
google-api-python-client==2.121.0
pydrive2==1.19.0